Move TtRssUrlReader into a separate file
