+++ mk/mk.deps
@@ -104,9 +104,10 @@ src/controller.o: src/controller.cpp include/controller.h include/cache.h \
  include/oldreaderurlreader.h include/opmlurlreader.h \
  include/regexmanager.h include/remoteapi.h include/rssfeed.h \
  include/utils.h include/rssparser.h include/stflpp.h include/strprintf.h \
- include/ttrssapi.h 3rd-party/json.hpp include/utils.h include/view.h \
- include/controller.h include/filebrowserformaction.h \
- include/formaction.h include/history.h include/keymap.h include/stflpp.h
+ include/ttrssapi.h 3rd-party/json.hpp include/ttrssurlreader.h \
+ include/utils.h include/view.h include/controller.h \
+ include/filebrowserformaction.h include/formaction.h include/history.h \
+ include/keymap.h include/stflpp.h
 src/dialogsformaction.o: src/dialogsformaction.cpp \
  include/dialogsformaction.h include/formaction.h include/history.h \
  include/keymap.h include/configparser.h include/configactionhandler.h \
@@ -453,7 +454,7 @@ src/rssparser.o: src/rssparser.cpp include/rssparser.h \
  include/remoteapi.h rss/feed.h rss/rssparser.h include/rssfeed.h \
  include/matchable.h include/rssitem.h include/utils.h include/logger.h \
  include/rssignores.h include/strprintf.h include/ttrssapi.h \
- 3rd-party/json.hpp include/cache.h include/urlreader.h include/utils.h
+ 3rd-party/json.hpp include/cache.h include/utils.h
 src/selectformaction.o: src/selectformaction.cpp \
  include/selectformaction.h include/filtercontainer.h \
  include/configparser.h include/configactionhandler.h \
@@ -487,15 +488,14 @@ src/textformatter.o: src/textformatter.cpp include/textformatter.h \
  include/logger.h config.h include/strprintf.h
 src/ttrssapi.o: src/ttrssapi.cpp include/ttrssapi.h 3rd-party/json.hpp \
  include/cache.h include/configcontainer.h include/configparser.h \
- include/configactionhandler.h include/remoteapi.h include/urlreader.h \
- include/logger.h config.h include/strprintf.h include/remoteapi.h \
- rss/feed.h rss/item.h include/strprintf.h include/utils.h \
- include/logger.h
-src/ttrssurlreader.o: src/ttrssurlreader.cpp include/ttrssapi.h \
- 3rd-party/json.hpp include/cache.h include/configcontainer.h \
- include/configparser.h include/configactionhandler.h include/remoteapi.h \
- include/urlreader.h include/fileurlreader.h include/logger.h config.h \
+ include/configactionhandler.h include/remoteapi.h include/logger.h \
+ config.h include/strprintf.h include/remoteapi.h rss/feed.h rss/item.h \
  include/strprintf.h include/utils.h include/logger.h
+src/ttrssurlreader.o: src/ttrssurlreader.cpp include/ttrssurlreader.h \
+ include/urlreader.h include/fileurlreader.h include/logger.h config.h \
+ include/strprintf.h include/remoteapi.h include/configcontainer.h \
+ include/configparser.h include/configactionhandler.h include/utils.h \
+ include/logger.h
 src/urlreader.o: src/urlreader.cpp include/urlreader.h
 src/urlviewformaction.o: src/urlviewformaction.cpp \
  include/urlviewformaction.h include/formaction.h include/history.h \
